requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
openai==0.28.0
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
textstat==0.7.3
sentence-transformers==2.2.2
python-Levenshtein==0.21.1
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0